+++ crates/icp/src/lib.rs
@@ -396,7 +396,7 @@ impl MockProjectLoader {
             configuration: Configuration::Connected {
                 connected: Connected {
                     api_url: "https://icp-api.io".parse().unwrap(),
-                    gateway_url: Some("https://icp0.io".parse().unwrap()),
+                    http_gateway_url: Some("https://icp0.io".parse().unwrap()),
                     root_key: None,
                 },
             },