+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Unreleased
 
-* feat: Optionally split connected networks' `url` into `api-url` and `gateway-url`
+* feat: Optionally split connected networks' `url` into `api-url` and `http-gateway-url`
 * feat: Allow specifying a version of the network launcher to use
 * feat: Support subaccounts and ICRC-1 IDs in `icp token`, `icp cycles`, and `icp identity account-id`
 * feat!: Recipes are now specified `@registry/recipe@version`, the version component is required. The `latest` version is no longer assumed and the version tags will be removed soon.